# swagger_aula.yml
openapi: 3.0.0
info:
  title: API de Mensagens
  description: API para gerenciar mensagens
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Servidor local
paths:
  /api/data:
    post:
      summary: Salva uma nova mensagem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MensagemInput'
      responses:
        '201':
          description: Mensagem criada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mensagem'
        '400':
          description: Requisição inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/data/{id}:
    get:
      summary: Obtém uma mensagem específica
      parameters:
        - in: path
          name: id
          required: true    
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mensagem'
        '404':
          description: Mensagem não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Atualiza uma mensagem existente
      parameters:
        - in: path
          name: id
          required: true    
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MensagemInput'
      responses:
        '200':
          description: Mensagem atualizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mensagem'
        '404':
          description: Mensagem não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Deleta uma mensagem
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Mensagem deletada com sucesso (sem conteúdo)
        '404':
          description: Mensagem não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'          
    patch:
      summary: Atualiza parcialmente uma mensagem
      parameters:
        - in: path
          name: id
          required: true    
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MensagemInput'
      responses:
        '200':
          description: Mensagem atualizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mensagem'
        '404':
          description: Mensagem não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Mensagem:
      type: object
      properties:
        id:
          type: integer
          format: int64
        message:
          type: string
    MensagemInput:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    Error:
      type: object
      properties:
        error:
          type: string